Make some changes to entities to match database types

+++ src/main/java/org/eastway/echarts/shared/Appointment.java
@@ -19,9 +19,9 @@ import java.util.Date;
 
 public interface Appointment {
 
-	public void setId(long id);
+	public void setId(int id);
 
-	public long getId();
+	public int getId();
 
 	public void setAppointmentDate(Date appointmentDate);
